+++ examples/copyContentToProjectFolder/applications/example/app/examples/mongodb/insertData.json/get.js
@@ -1,0 +1,6 @@
+module.exports = function(client, callback) {
+	dbAlias.testCollection.insert(client.query, function(err) {
+		client.context.data = !err;
+		callback();
+	});
+}