+++ examples/copyContentToProjectFolder/applications/example/app/examples/mysql/getCities.json/get.js
@@ -1,6 +1,5 @@
 module.exports = function(client, callback) {
-	db.polltool.query('select * from City', function(err, rows, fields) {
-		if (err) throw err;
+	dbAliad.query('select * from City', function(err, rows, fields) {
 		client.context.data = { rows:rows, fields:fields };
 		callback();
 	});