+++ examples/copyContentToProjectFolder/applications/example/app/examples/security/getUsers.json/get.js
@@ -1,5 +1,4 @@
 module.exports = function(client, callback) {
-	client.context.data = [];
 	impress.security.db.users.find({}).toArray(function(err, nodes) {
 		client.context.data = nodes;
 		callback();